Added function to bootstrap new python projects
This allows to bootstrap python projects by using the function
pyconf-bootstrap-pyproject. Given a target dir, it will create a
new pyenv for the project in target dir. It's also possible to install
all dependencies using poetry and define a set of dependencies to
install in all projects (like black, debugpy, etc...)