feat: Added bootstrap transient menus, remote exec support
Added transient menus to the bootstrap feature. Now new projects can
be bootstrapped using with a user friendlier interface.
Added remote execution in docker containers. This is for docker
containers that are running on local machine only for now.
Refactored menu parameters to be decoupled.
Refactored some of the functions to share functionality between
transient and non-transient interactive functions.